Extract tun_listener and ip_packet_router
